+++ SalaryComparer/Program.cs
@@ -1,42 +1,82 @@
-﻿using System;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using Microsoft.Extensions.Configuration;
 
 namespace SalaryComparer
 {
     class Program
     {
+        private static IConfigurationRoot _configuration;
+
         static void Main(string[] args)
         {
-            var current = new Salary(double.Parse(args[0]), double.Parse(args[1]), double.Parse(args[2]), true);
-            var proposed = new Salary(double.Parse(args[3]), double.Parse(args[4]), double.Parse(args[5]), true);
+            _configuration = new ConfigurationBuilder()
+                .SetBasePath(Directory.GetParent(AppContext.BaseDirectory).FullName)
+                .AddJsonFile("appsettings.json", optional: false)
+                .Build();
 
-            WriteToConsole($@"
-                              |   Current   |  Proposed   | Difference  |
-------------------------------|-------------|-------------|-------------|");
-            WriteCurrencyLine("Salary                       ", new[] { current.Amount, proposed.Amount, proposed.Amount - current.Amount });
-            WriteCurrencyLine("Pension Contribution         ", new[] {
-current.EmployeeContribution.Amount, proposed.EmployeeContribution.Amount, proposed.EmployeeContribution.Amount - current.EmployeeContribution.Amount });
-            WriteCurrencyLine("Employer Pension Contribution", new[] { current.EmployerContribution.Amount, proposed.EmployerContribution.Amount, proposed.EmployerContribution.Amount - current.EmployerContribution.Amount });
-            WriteCurrencyLine("Income Tax                   ", new[] { current.TotalIncomeTax, proposed.TotalIncomeTax, proposed.TotalIncomeTax - current.TotalIncomeTax });
-            WriteCurrencyLine("National Insurance           ", new[] { current.TotalNationalInsurance, proposed.TotalNationalInsurance, proposed.TotalNationalInsurance - current.TotalNationalInsurance });
-            if (current.StudentLoan || proposed.StudentLoan)
+            var salariesConfiguration = _configuration.GetSection("Salaries").GetChildren();
+            var salaries = new List<Salary>();
+            foreach (var salary in salariesConfiguration)
             {
-                WriteCurrencyLine("Student Loan                 ", new[] { current.TotalStudentLoan, proposed.TotalStudentLoan, proposed.TotalStudentLoan - current.TotalStudentLoan });
+                var name = salary.GetSection("Name").Value.ToString();
+                var amount = double.Parse(salary.GetSection("Amount").Value);
+                var employeePensionContribution = double.Parse(salary.GetSection("PensionContributionPercentage:Employee").Value);
+                var employerPensionContribution = double.Parse(salary.GetSection("PensionContributionPercentage:Employer").Value);
+                var studentLoan = bool.Parse(salary.GetSection("StudentLoan").Value);
+
+                salaries.Add(new Salary(name, amount, employeePensionContribution, employerPensionContribution, studentLoan));
             }
 
-            WriteToConsole("------------------------------|-------------|-------------|-------------|");
-            WriteCurrencyLine("Total Deductions             ", new[] { current.TotalDeductions, proposed.TotalDeductions, proposed.TotalDeductions - current.TotalDeductions });
-            WriteCurrencyLine("Total Pension Contribution   ", new[] { current.TotalPensionContribution, proposed.TotalPensionContribution, proposed.TotalPensionContribution - current.TotalPensionContribution });
-            WriteCurrencyLine("Take Home                    ", new[] { current.TakeHome, proposed.TakeHome, proposed.TakeHome - current.TakeHome });
-            WriteToConsole("------------------------------|-------------|-------------|-------------|");
-            WriteCurrencyLine("Net Gain                     ", new[] { current.NetGain, proposed.NetGain, proposed.NetGain - current.NetGain });
+            PrintSummary(salaries);
         }
 
-        private static void WriteCurrencyLine(string lineTitle, double[] items, string separator = " | ")
+        private static void PrintSummary(IList<Salary> salaries)
         {
-            var output = $"{lineTitle}{separator}";
+            var separator = "|";
+            var showDiff = salaries.Count == 2;
+
+            var heading = $"                               {separator}";
+            foreach (var salary in salaries)
+            {
+                heading += $" {salary.Name,-11} {separator}";
+            }
+
+            WriteToConsole(heading);
+            WriteDividingLine(salaries.Count, separator);
+            WriteCurrencyLine("Salary", salaries.Select(x => x.Amount).ToArray(), separator);
+            WriteCurrencyLine("Pension Contribution", salaries.Select(x => x.EmployeeContribution.Amount).ToArray(), separator);
+            WriteCurrencyLine("Employer Pension Contribution", salaries.Select(x => x.EmployerContribution.Amount).ToArray(), separator);
+            WriteCurrencyLine("Income Tax", salaries.Select(x => x.TotalIncomeTax).ToArray(), separator);
+            WriteCurrencyLine("National Insurance", salaries.Select(x => x.TotalNationalInsurance).ToArray(), separator);
+            WriteCurrencyLine("Student Loan", salaries.Select(x => x.TotalStudentLoan).ToArray(), separator);
+            WriteDividingLine(salaries.Count, separator);
+            WriteCurrencyLine("Total Deductions", salaries.Select(x => x.TotalDeductions).ToArray(), separator);
+            WriteCurrencyLine("Total Pension Contribution", salaries.Select(x => x.TotalPensionContribution).ToArray(), separator);
+            WriteCurrencyLine("Take Home", salaries.Select(x => x.TakeHome).ToArray(), separator);
+            WriteDividingLine(salaries.Count, separator);
+            WriteCurrencyLine("Net Gain", salaries.Select(x => x.NetGain).ToArray(), separator);
+        }
+
+        private static void WriteCurrencyLine(string lineTitle, double[] items, string separator = "|")
+        {
+            var output = $"{lineTitle,-30} {separator}";
             foreach (var item in items)
             {
-                output += $"{item,11:C}{separator}";
+                output += $" {item,11:C} {separator}";
+            }
+
+            WriteToConsole(output);
+        }
+
+        private static void WriteDividingLine(int columns, string separator = "|")
+        {
+            var output = $"-------------------------------{separator}";
+            for (int i = 0; i < columns; i++)
+            {
+                output += $"-------------{separator}";
             }
 
             WriteToConsole(output);