+++ build.docker.ps1
@@ -7,7 +7,7 @@ if ($LASTEXITCODE -ne 0) {
     exit $LASTEXITCODE
 }
 
-docker run --rm salary-comparer $args
+docker run --rm salary-comparer # $args
 if ($LASTEXITCODE -ne 0) {
     exit $LASTEXITCODE
 }